Refactoring UITestExecutor into two parts
